Add parameter checking for login command
This change introduces parameter checking for the login command.
If the parameters are not as required, an error message is printed
to STDERR and the exit status is set to 1.

The checks performed are:
* Ensure that at least one parameter (the URL) is provided
* Ensure that a password is provided if the username is provided

This implementation makes use of panic and recover within the
commands.(*Login).Run function.  However, it may be better to use
a globally defined error reporting function, to ensure consistency
of handling for all of the commands.