Amend scopes to permit use of scope token as a command.
Currently the scope tokens (application, entity, effector, config, activity)
can't be used as commands, as they are interpreted as scopes.

This change amends the condition (scope.go:83) that determines whether a token
is a scope or a command, to interpret it as a scope ONLY IF there are AT LEAST
two subsequent parameters - one for the id of the scope, and one command.

This allows us to have a command, with one argument, that has the same name
as a scope token.  We can now use this to allow such commands to be used as
"show" commands.

Included in this change is the restoration of the application command, which
was renamed as 'status' in the previous commit, thus:

$ br application myweb
Name    Id         Status    Location
myweb   Ij5EHejA   RUNNING   WQmFQ3SG