Improve the handling of sub-commands.
This is done so that it's easier to define and use subcommands.  At the moment
we don't need any such commands, but before we remove "list" as a supercommand
it is worth putting this support in place.  For how to use it see list.go as
an example and the use of it that is made in factory.go:44.