Introduce a generic Exit/Error handler and use this for the login command
The error_handler package provides function ErrorExit; the function prints
out an error message to STDERR and exits with a given or default exit code.
The first argument can be of type string or error.  The second argument
(optional) should be of type integer and is for setting the exit code.
Three integer constants are defined for setting the exit code:
  CLIUsageErrorExitCode - for reporting command syntax errors
  CliGenericErrorExitCode - for reporting processing errors
  CLITrapErrorCode - for reporting abnormal/trapped errors
If the second argument is not provided the CliGenericErrorExitCode value is used.

The login command has been extended to connect to Brooklyn and obtain the version
string which is then printed to STDOUT.