Update README for runtime.
Splits README.md into two, one file still called README.md for developers who wish to get and build the project themselves, and one called simply README with information for users getting a built copy of the tool.

Adds the README to the Maven artifact instead of the README.md.